+++ cadwyn/routing.py
@@ -43,7 +43,7 @@ from starlette.routing import (
 from typing_extensions import Self, assert_never
 from verselect.routing import VERSION_HEADER_FORMAT
 
-from cadwyn._compat import PYDANTIC_V2
+from cadwyn._compat import model_fields, rebuild_fastapi_body_param
 from cadwyn._utils import Sentinel, UnionType, get_another_version_of_module
 from cadwyn.codegen import _get_package_path_from_module, _get_version_dir_path
 from cadwyn.exceptions import CadwynError, ModuleIsNotVersionedError, RouteAlreadyExistsError, RouterGenerationError
@@ -85,7 +85,7 @@ class _RouterInfo(Generic[_R]):
 
 class InternalRepresentationOf:
     def __class_getitem__(cls, original_schema: type, /) -> type[Self]:
-        return type("InternalBodyRequest", (cls, original_schema), {})
+        return cast(Any, type("InternalRepresentationOf", (cls, original_schema), {}))
 
 
 def generate_versioned_routers(
@@ -169,22 +169,7 @@ class _EndpointTransformer(Generic[_R]):
                 body_param: FastAPIModelField = copy_of_dependant.body_params[0]
                 body_schema = body_param.type_
                 new_type = schema_to_internal_request_body_representation.get(body_schema, body_schema)
-                kwargs: dict[str, Any] = {"name": body_param.name, "field_info": body_param.field_info}
-                if PYDANTIC_V2:
-                    body_param.field_info.annotation = new_type
-                    kwargs.update({"mode": body_param.mode})
-                else:
-                    kwargs.update(
-                        {
-                            "type_": new_type,
-                            "class_validators": body_param.class_validators,
-                            "default": body_param.default,
-                            "required": body_param.required,
-                            "model_config": body_param.model_config,
-                            "alias": body_param.alias,
-                        },
-                    )
-                new_body_param = FastAPIModelField(**kwargs)
+                new_body_param = rebuild_fastapi_body_param(body_param, new_type)
                 copy_of_dependant.body_params = [new_body_param]
 
             for older_router_info in list(router_infos.values()):
@@ -483,7 +468,7 @@ class _AnnotationTransformer:
         elif isinstance(annotation, type):
             if annotation.__module__ == "pydantic.main" and issubclass(annotation, BaseModel):
                 return create_body_model(
-                    fields=self._change_version_of_annotations(annotation.__fields__, version_dir).values(),
+                    fields=self._change_version_of_annotations(model_fields(annotation), version_dir).values(),
                     model_name=annotation.__name__,
                 )
             return self._change_version_of_type(annotation, version_dir)
@@ -646,6 +631,10 @@ def _generate_signature(
     default_counter = 0
     for param in old_params.values():
         if param.default is not inspect.Signature.empty:
+            assert new_callable.__defaults__ is not None, (  # noqa: S101
+                "Defaults cannot be None here. If it is, you have found a bug in Cadwyn. "
+                "Please, report it in our issue tracker."
+            )
             default = new_callable.__defaults__[default_counter]
             default_counter += 1
         else: