+++ CHANGELOG.md
@@ -28,6 +28,10 @@ Please follow [the Keep a Changelog standard](https://keepachangelog.com/en/1.0.
 * `latest_schemas_module` was renamed to `latest_schemas_package` everywhere
 * `api_version_var` in `VersionBundle` is now an optional argument instead of a required one
 
+### Removed
+
+* `cadwyn.internal_representation_of` because it is now done using [an annotation](docs.cadwyn.org/reference#internal-request-body-representations)
+
 ## [2.3.4]
 
 ### Fixed